Remove initialized api from tensor
Summary: In aten tensor this is used to reason about if the Intrusive ptr to tensorImpl is set or not. We were using it to reason about the data ptr so I'd rather just delete it. If we need to reason about if the TensorImpl ptr in Tensor is null or not we can add it back with different semantics.

Reviewed By: larryliu0820

Differential Revision: D46129275

fbshipit-source-id: 0ea311b1e5b2ae6048eabe20dbc7afcd41c0fe7d