Fix External ID Order
Summary:
In the case where we encounter the second argument before the first argument such as this:

```
def forward(self, x, y):
    return torch.cat((y, x))
```

The external IDs are in order of x:0, y:1 but the internal ids are x:1, y:0 since we encounter y first.

Generally, in the case where we serialize external_id != internal_id, then the order for which we add the ids into our executor class will be unsorted. This causes issues as when we prepare the inputs for execution, we wire them to the wrong inputs for the xnnpack graph.

Reviewed By: digantdesai

Differential Revision: D47280248

fbshipit-source-id: a0657bfe3b4fb8c0aeca09edd2fcf4ffd08ff04f