Conv with padding = "valid"
Summary:
Bug with lowering torch.nn.conv2d when padding is 1 dimensional.

Atttributes like stride, kernel_size, etc. can be supplied with single ints. But in the captured graph they are represented as [single_int, single_int]. This is not the case with padding. When we supply a single int or valid, the captured graph shows it as [single_int]. Thus, we must fix for this case so that we correctly lower the operator

Reviewed By: digantdesai

Differential Revision: D47488293

fbshipit-source-id: 49d949284f0a17729ed7fa8fa9dbb01be4470ac0