Move executor_runner under /sdk/runners/
Summary:
Over time, `executor_runner.cpp` has evolved from a simple example to becoming the commandline that we use to drive different kinds of testing. This is fine, since we need a tool like this, but it's no longer a good initial example for new users, since it does things that they shouldn't. It also depends on non-core pieces that will eventually live in a different repo, like all of the optimized kernels and some custom ASR ops.

Since it's closely tied to SDK concepts like etdump and bundled_program, move it under `/sdk/`.

Reviewed By: larryliu0820

Differential Revision: D47504325

fbshipit-source-id: 8a26e79870557448806e234d415ae401a0be2d4e