Fix delegates
Summary:
With Yidi's migration in D46729844, the typing for the delegates is incorrect/inconsistent, with to_backend and preprocess taking in ExirExportedProgram sometimes, and other times just taking in an fx.GraphModule (pyre-strict is not enabled for the backend/ folder, so it is not caught by typechecking). This diff updates the delegate APIs to take in ExportedProgram.

https://docs.google.com/document/d/1vpvjnGk1TWdnVbgrzuTBJTOFx2D06GTaMoZIrSipWAc/edit

Reviewed By: mergennachin

Differential Revision: D47252888

fbshipit-source-id: 647e26d49dcafd2d08b1a91c1565515c31159ef8