Fix Slice Memory
Summary:
There is a bug with some of the tensor manipulation functions when we operate them in NHWC. We serialize the permuted tensor shape. But other arguemtns which reference dimensions within the shape are not properly permuted as well.

For example, if we are slicing dim 1. If we add a convolution before this node our memory format tagging will tag this node to operate in NHWC. As a result, we are actually slicing dim 3 because the tensor was permuted.

Reviewed By: digantdesai

Differential Revision: D47413374

fbshipit-source-id: d6bd661a30d084a48756a9f06fc5d8d16e12aebd