Fix Concat Memory
Summary:
There is a bug with some of the tensor manipulation functions when we operate them in NHWC. We serialize the permuted tensor shape. But other arguments which reference dimensions within the shape are not properly permuted as well.

For example, if we are concating two tensors along dim 1. If we add a convolution before this node our memory format tagging will tag this node to operate in NHWC. As a result, we will actually be concatenating against dim 3 because the tensor was permuted.

Reviewed By: digantdesai

Differential Revision: D47413372

fbshipit-source-id: 3a1cd9b0ca678d2d673fbb09ca2a35888d61f986