Remove dependency on pt_operator_library
Summary:
In `codegen.bzl` there's this logic to reuse `pt_operator_library` in `et_operator_library` just so we can save some code.

This is not playing well with OSS plan, since we don't have access to `pt_operator_library` in OSS executorch. This diff adds the missing feature to `et_operator_library` which is to support `include_all_operators`.

Reviewed By: dbort

Differential Revision: D47571115

fbshipit-source-id: ea8ef7a8571eb19db0b0a50787c55a322bd29230