Move executorch_prim_ops_registry into exir
Summary: We need prim ops to be registered at exir level to support sym ops. For some reason this piece of code currently lives in executorch/kernels. With the oss plan we need to build package for exir and this target needs to be part of the package. This diff moves the target into exir.

Reviewed By: JacobSzwejbka

Differential Revision: D47646479

fbshipit-source-id: 378a1751675218920ddb9d87eb481c8cba9665cc