Make codegen.bzl oss compliant
Summary:
This is part of an effort to enable OSS BUCK2 build for Executorch.

codegen.bzl is required for kernel registration and it's currently accessing external (non-executorch) deps. This diff uses the new `external_deps` feature added to `runtime_wrapper` to support these deps.

Notice that we have two places depending on `libtorch` in fbcode but we depend on `torch_mobile_all_ops` and `torch` respectively in fbsource. Here I'm choosing  `torch_mobile_all_ops` and see if it passes CI.

Reviewed By: cccclai, dbort

Differential Revision: D47615228

fbshipit-source-id: 86716736958f605a5e377f18e81e3ef4be94d99a