Refactor exir/__init__.py
Summary: We had been just dumping everything in `__init__.py` which hurts readibility and makes it difficult to reason about what symbols we are exposing as public. This is the first pass effort to try and move things out. Im not 100% sold on capture/_config.py but I think its moving in the right direction which is good enough for me with this diff.

Reviewed By: angelayi

Differential Revision: D47604227

fbshipit-source-id: 8dd7c6d36d6992d0f09245d3d7fb624fce22dee0