Fix codegen/tools targets
Summary:
Move the following targets into `targets.bzl`, so they can be shared by oss and internal:

* gen_oplist_lib
* gen_all_oplist_lib
* gen_all_oplist
* test_gen_all_oplist

The following targets used to live in `TARGETS` but they are not runnable in oss yet. The workaround here I'm introducing is to guard them by `is_oss()` and a boolean passed in from `TARGETS`. This makes sure they only appear in fbcode.

* gen_functions_yaml
* test_gen_oplist_real_model

Also notice the current `external_deps` patching doesn't work well with `select` in `deps` so have to workaround it.

Reviewed By: dbort

Differential Revision: D47661541

fbshipit-source-id: a45c13b02c679d2ab8b4f9281bc17fa8395999d1