Remove profiler time hooks deps
Summary: This diff gets rid of the need for separate time hooks in the profiler code base. We switch over the profiler to use the time hook available in PAL. For Jarvis added an implementation for `et_pal_current_ticks` that is only brought in when compiling for Xtensa. If compiling for devserver in Jarvis, it'll fall back to the default POSIX implementation.

Reviewed By: larryliu0820, cccclai

Differential Revision: D47637865

fbshipit-source-id: 86c61c7f00bfa9be18c8257c115ecb1d0bb754a3