Move QuantFusionPass out of EdgeCompileConfig
Summary:
By definition `QuantFusionPass` is replacing some pattern in edge dialect to ops in `quantized_decomposed` namespace. According to this definition, `QuantFusionPass` should be a backend pass because it introduces `quantized_decomposed` ops which belong to backend ops. Hence we want to move it out of `EdgeCompileConfig` and recommend users to add it in `ExecutorchBackendConfig.passes`.

Also removing `SpecPropPass()` in the test since it's already added by default.

Reviewed By: angelayi

Differential Revision: D47731943

fbshipit-source-id: a138ad9bee1e43b4949ca3c299961adb4785a667