Update Stage interace to have graph_module prop
Summary: More flexible for inspection than a str. For example, if we want to do some detailed analysis on a stage's artifact, an interface to get a graph_module in addition to artifact itself is helpful. Getting a str from this should be uniform, and trivial. Except for Serialize where we don't have either anyway.

Reviewed By: mcr229

Differential Revision: D47775454

fbshipit-source-id: 07c12ac069fb25bbf28cb650225037ab576c62ff