Move SpecPropPass after backend passes
Summary:
SpecPropPass is used to construct the `meta["spec"]` metadata which is later used for memory planning. Specifically, it stores the shapes/dtypes used for memory planning (symbolic shapes are converted to their upper bound value in the SymShapeEvalPass), and in the memory planning pass, it stores the lifetimes of tensors and their buffer sizes.

This diff moves SpecPropPass to after user-defined passes are run so that user passes do not affect the `spec` metadata (ex. If users insert a node then they would need to update meta["spec"] because we do not recalculate it). Originally this was put before user-defined passes because Turing team has a hacky pass to change `spec.dtype` due to the CPU kernel not matching the hardware kernel, but from a high level I think this should still go at the end.

Turing team will fix + remove their "legalize_dtype" pass in D47640453 so that it is not dependent on meta["spec"].

Reviewed By: zhxchen17

Differential Revision: D47572569

fbshipit-source-id: efb77df1c867925c0b22a53d8435dae515e033c8