Rename ExecutionPlan to Method
Summary:
The concept of an "execution plan" was always meant to be hidden from users. Its purpose has also changed a bit over time. In the beginning, it was supposed to be a way to let a single program file run on multiple target systems. But now it represents a method or entry point of the model.

Rename it, but leave the old name around for now until the new APIs are ready and everyone has moved off of the old name.

Reviewed By: tarun292, JacobSzwejbka

Differential Revision: D47688621

fbshipit-source-id: a842d7a672cb27a32281917ab48efee3f8555b5a