arrange created submodule args
Summary:
**The goal of this diff is for the submodule produced by `create_submodule_from_nodes` to have the correct ordering of arguments**

**Reasoning**
We are migrating to torch.export for graphs with lifted params. As a result we plan on delegates to consume lifted exported programs. When we partition nodes from the graph of the original exported program we must create a new ExportedProgram from them for delegate consumpation. In order to create a correct ExportedProgram for preprocess, we need the submodule arguments to be in the following order:

`(params + buffers)(user args)`

that is the user args must be the last few arguments.

**Current State**
Currently, the way in which we fuse partitioned node lists together into a graph module is in topological sort with no knowledge of which args are parameters and which args are user inputs. As a result, the graph module's placeholder arguments can be in any such order.

Reviewed By: angelayi, cccclai

Differential Revision: D47783522

fbshipit-source-id: a9b953f0d45cc971e1ac81bd71886285eaacd5c2