Add sleef as a direct fbandroid_platform_deps of operator targets
Summary:
Seems that `fbandroid_platform_deps` are not transitive between dependencies, so optimized operator targets that depend on libvec (and therefore sleef) may experience a linking error with sleef when building for android platforms.

The solution is to add sleef as a direct `fbandroid_platform_deps` of the operator target.

This should fix the issue described in this [workplace post](https://fb.workplace.com/groups/pytorch.edge.users/permalink/1357728431764113/).

Reviewed By: kirklandsign

Differential Revision: D47974433

fbshipit-source-id: 1cbf1d34a3dc1d83c89508c652bf412badfcc8bf