extract_sources.py generates cmake sources lists from buck2
Summary:
The new `extract_sources.py` tool can query and filter the source files associated with sets of buck2 targets, and write them out to a file that can by included by CMake. In the future, the tool could also support other formats like Make or Ninja.

This will make it much easier to keep a parallel CMake-based system in sync with the source of truth in buck2.

Reviewed By: larryliu0820

Differential Revision: D47979069

fbshipit-source-id: f2ad6b52a34211878c2dd2d5c2b06b62dde73353