map third party deps between OSS and Internal
Summary:
In order for the XNNPACK delegate's buck targets to work both internally and for OSS, we need to be able to switch the third-party XNNPACK dependency depending on whether or not we are building for OSS or for Internal use.

We have something similar in PyTorch, so we are doing something similar here. We disentangle this from ExecuTorch runtime wrapper in order to separate the XNNPACK delegate's third-party dependency from Core Executorch's Third-Party Dependencies.

Reviewed By: dbort

Differential Revision: D48214072

fbshipit-source-id: f12505f6b96a4bf7066a46f2f9035cc4ce64f3b6