Bump flatbuffer version
Summary:
As titled, bump flatbuffer version from v12 to v23.5.

In flatbuffer v23.5 the default alignment became 32 and it may cause some issue because it starts to diverge from the fbcode flatbuffer version v12.

To close the gap, in `program.cpp` we want to check if the flatbuffer is aligned with the max alignment (the size of `long double` according to https://en.cppreference.com/w/cpp/types/max_align_t) instead of following `FLATBUFFERS_MAX_ALIGNMENT`(which is 16 because internal flatbuffer version is lacking behind).

Reviewed By: dbort

Differential Revision: D48325610

fbshipit-source-id: d45febf468ad99fe516e2ed86181dd922de899c1