Fix ScalarType -> C type conversion in quantize_per_tensor
Summary:
The `quantize_per_tensor` implementation was using the
wrong type to assert that the lower bound was correct,
it was using `char` (which can be signed or unsigned depending
on the platform) instead of `int8_t` which is always signed.

`signed char` would also work fine, but I figured we should
use the C types already declared in the macros.

Reviewed By: JacobSzwejbka

Differential Revision: D48415930

fbshipit-source-id: 090a1f4d2c9824daeaae7c5bb5f022bbaef77131