Add SymIntToTensorPass in to_edge
Summary:
JacobSzwejbka ran into some issues when lowering to executorch where we
have the following graph:
```
%symfloat_1 = executorch_prim::mul(%symfloat_a, %symfloat_b)
%mul_out = aten::mul.out(%tensor, %symfloat_1, %mul_out, %mul_out)
```
Which errors because mul.out is being passed a symfloat, but mul.out only
accepts 2 tensors. So if we add this pass, the symfloats will be converted to
tensors.

Apparently I wrote this pass back in April but never put it somewhere where it
gets run by default.

Reviewed By: JacobSzwejbka

Differential Revision: D48415403

fbshipit-source-id: 1ce9db145feb8b54a666e38e4f7aefdb1e314a48