Fix linear's with permute copy (#79)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/79

There are some issues with permute_copy in both partitioner as well as convert_to_linear pass

### Partitioner:
For Quantized Partitions, we fail to pull in the q/dq nodes above permute_copy.

```
get_attr --> q --> dq --> permute_copy --> addm
```
The solution is checking the inputs to the source_partition for permute_copy node, and if it is one of them, then we add it to the partition and check its inputs

### Convert to Linear Pass
In the pattern
```
get_attr --> q --> dq --> permute_copy --> addmm
```
We replace the entire source partition with just linear, however we fail to delete the permute_copy because it is an input to the source partition instead of the dq (dq should actually be the input to the linear source partition). The weight given to linear should not be the result of permute copy, but should instead be the input to permute copy.

This happens because q and dq are not tagged as part of the linear source partition, so permute_copy becomes the input to the source partition.

Differential Revision: D48488931

fbshipit-source-id: 6269660d4c07cb9d3a7c5c020efd8ce753a40fe3