Update toplevel graph signature (#95)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/95

After partitioning the graph for delegation, the graph signature of
the toplevel graph may change due. Specifically, a change that was seen was
that the output nodes of the graph are different. So, I refactored a function
to always update the graph signature after to_backend's partitioning API.

We (export) should figure out a way so that we don't have to do this manual
fixing of the graph signature. Probably retracing is the correct solution...but
it will break a couple of stuff right now, specifically the metadata.

Reviewed By: gmagogsfm

Differential Revision: D48529727

fbshipit-source-id: 96cb9d567b78c89bf27a4173538d310c0c45eb51