Add schema version to serializer/deserializer (#107420)
Summary:
Added a version number to the schema for BC issues. We will add this number to the serialized ExportedProgram and then when deserializing, if the number does not match up with the existing deserializer, we will error. We should update the number of there are any major changes to the schema.

X-link: https://github.com/pytorch/pytorch/pull/107420
Approved by: https://github.com/zhxchen17

Reviewed By: gmagogsfm, tugsbayasgalan

Differential Revision: D48549362

Pulled By: angelayi

fbshipit-source-id: 83a67bcdfeea2d2d05bf27283f46fe1f6bc12ad5