Introduce a new interface in lib.py and the ETInspector class (#111)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/111

## About this stack
To upgrade the SDK debugger so that it offers not only the CLI experience, but also public APIs that return DataFrames containing model architecture info and profiling stats info.
Once this tool is in a usable state, we can migrate the new logic to the internal flow (executorch/sdk/fb/lib.py)

## About this diff
Define the skeleton of the ETInspector class (including all the high-pri APIs). In lib.py, defines the APIs that initialize an ETInspector.
Note that this diff basically turns the CLI useless temporarily. No one is using the CLI rn so it should be fine.

## Next
Define the DataFrame schemas

Reviewed By: Jack-Khuu

Differential Revision: D48456830

fbshipit-source-id: 7ac6327da3ab07e273ca14667a5e6fadd3407ff6