fix permute copy name (#80)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/80

We have a node_visitor for permute already, it was mistakenly named static_transpose after its xnnpack counterpart. permute_copy should not be skipped. It should only be skipped when applied to linear weights.

Instead our linear conversion pass will convert away any permute_copies which are not actually to be executed.

Reviewed By: kirklandsign

Differential Revision: D48488930

fbshipit-source-id: 9d84e00bdfa04828768e8d0e352817a0348b0cc6