Remove strange helpers for python -> tensor conversions, and at -> et tensor conversions (#121)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/121

Keepalive was difficult to manage and Im not even sure how sound it was for things like strides and dim order. Now the run_method loop is a little grosser, but I think the intent is much clearer. We cast from python types to at types, and then possible have to do a hard conversion from at::tensor to ETensor which can be pretty gross since ETensor doesnt manage any memory. This highlights the need for like an ETensor wrapper that manages a bunch of state like sizes and TensorImpl

Reviewed By: dbort

Differential Revision: D48618092

fbshipit-source-id: cea1feae3f801be1c6857d148a48335f9408dca0