Make kernels/portable/cpu compatible with C++11/14/17/20
Summary:
Tweak the portable kernels to build cleanly with C++11 as well as with more recent versions. This only covers non-test, non-aten targets under `//executorch/kernels/portable/cpu/...`.

A lot of the template changes deal with the lack of <base>_t templates in C++11. Most of them are of the form
```
template<class T>
using BASE_t = typename BASE<T>::type;
```
so the fix was to inline that pattern: add `typename` to the front and `::type` to the end.

Differential Revision: D48668113

fbshipit-source-id: d7792ddcaccdcdbce19534f49bc5db28761f9ee3