Use IDs for placeholder/outputs of implicit q/dq nodes
Summary:
In XNNPACK from the partitioner we encounter quantize node that are outputs to the graph. Since these quantize nodes are implicit and not actually quantizing the values, we must give its externalness to the node with which it quantizes.

### Outputs
```
                 |
conv --> q ----> | ----> dq -->
                 |
         Delegate boundary
```
In this example the output of the delegate is the output of quantized Conv. In XNNPACK, we must designate the output of Conv to be an external value. In this graph, however, the q node is the output. Since q is an implicit node (it does not actually running quantize) We pass its externalness to the conv node.

### Inputs
i.e.
```
        |
q ----> | ----> dq --> conv -->
        |
Delegate boundary
```
In this example q is the placeholder node and is given as input to the first dq. But we are running a quantized conv within the delegate. Since dq is implicit, we pass the externalness of the q place holder to dq, so that the input to Conv is external.

Reviewed By: digantdesai

Differential Revision: D48667676

fbshipit-source-id: a6adfdd917a0efd4de56549652900ea614cb9661