Fix XNNPACK Example for MV3 (#147)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/147

xnnpack requires dequant nodes to be duplicated such that len(dq.users) == 1. We run this pass on our internal XNNPACK tests

From Jerry: This will be a config that is used within the XNNPACKQuantizer. This will be added later, for the sake of enabling this in the example, we add the duplicatedequantnodepass to our edge compile config

Additionally, graph_validation is failing for MV3, so we have to disable_validation for now. cccclai is looking into it

Reviewed By: digantdesai

Differential Revision: D48703639

fbshipit-source-id: 4521243d120b401da47d048a10e1363b444b2cbb