Make kernels/portable/cpu compatible with C++11/14/17/20 (#128)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/128

Tweak the portable kernels to build cleanly with C++11 as well as with more recent versions. This only covers non-test, non-aten targets under `//executorch/kernels/portable/cpu/...`.

A lot of the template changes deal with the lack of <base>_t templates in C++11. Most of them are of the form
```
template<class T>
using BASE_t = typename BASE<T>::type;
```
so the fix was to inline that pattern: add `typename` to the front and `::type` to the end.

Reviewed By: digantdesai, kirklandsign

Differential Revision: D48668113

fbshipit-source-id: 056fe6706e55c3d52cc1cc67abf593f72c05a9b6