Add docs README.md and a custom directive (#208)
Summary:

Add a basic README for the documentation folder that gives a basic overview of how to build locally and contribute, as well as provide tips on how to author better pages with MyST parser.
Add a custom Sphinx directive that adds a code block with a nightly PyTorch version from the nightly.txt file.

Reviewed By: mergennachin, dbort

Differential Revision: D48757229