Add docs README.md and a custom directive (#208)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/208

Add a basic README for the documentation folder that gives a basic overview of how to build locally and contribute, as well as provide tips on how to author better pages with MyST parser.
Add a custom Sphinx directive that adds a code block with a nightly PyTorch version from the nightly.txt file.

Reviewed By: mergennachin, dbort

Differential Revision: D48757229

fbshipit-source-id: 6a83d46ae65a694bcc24bced88d4a999b998a94b