Adding EventTracer in runtime and support for it in Method (#232)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/232

EventTracer is a class that users can inherit and implement to log/serialize/stream etc. the profiling and debugging events that are generated at runtime for a model. An example of this is the ETDump implementation in the SDK codebase that serializes these events to a flatbuffer.

`ETDump` in the SDK folder will essentially be an implementation of this class serializing all this data to a flatbuffer.

Reviewed By: dbort

Differential Revision: D48743191

fbshipit-source-id: 290cd572c15286f3989bb22a2a1c45e3d5e99fb0