linearize_access_indexes only needs dim of broadcast_to (#270)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/270

The utility function `linearize_access_indexes` currently takes the `broadcast_to` tensor as the second argument. However, only the dimension of the tensor is needed.
Updating the function, to take only the dimension, so that it can be used in scenarios where a broadcast_to tensor is not available, and only a broadcast shape/ndim is known.

Reviewed By: kirklandsign

Differential Revision: D49159987

fbshipit-source-id: c6532b2a48b823234accdc3e69edc0e21147658a