Update benchmark results with android runs and compare to LI
Summary:
Add the benchmark run for Android; also compare the same setup (1 thread, 10 warmup, 50 runs) with LI to showcase the perf improvements.

Also drop the column for the first iteration. It's not predictable because the CPU frequency is not deterministic during warm-up.

Created from CodeHub with https://fburl.com/edit-in-codehub

Differential Revision: D49196911