Fix emitter bug which doesn't allow input lists to delegates (#282)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/282

As the delegate op doesn't have a schema we pass in `None` here when emitting the delegate arguments.

https://www.internalfb.com/code/fbsource/[4c61670ae0bd654ca630c2288f685d3a4dfa8ece]/fbcode/executorch/exir/emit/_emitter.py?lines=954

`emit_list` expects a `val_type` to be passed in though and triggers an assert when `None` is passed in.

https://www.internalfb.com/code/fbsource/[4c61670ae0bd654ca630c2288f685d3a4dfa8ece]/fbcode/executorch/exir/emit/_emitter.py?lines=471

To handle this we try to infer the `JitType` for lists using the newly added function `_get_list_jit_type`.

Reviewed By: cccclai

Differential Revision: D49180438

fbshipit-source-id: 44d6ab1fdaecdd74ec784d2b3e8e7b15fb63c9d4