rename to backend/interface.h (#285)
Summary:

this file is more about the interface between the backend and the executor runtime. The original name is `backend.h` and was renamed to `backend_registry.h` during refactor. But the backend registery is a really small part.

Techincally we can seperate out backend_registery, but it's just a few lines and feel like it might be overkill. Users also need to include two header files too and doesn't seem worth the effort

Reviewed By: JacobSzwejbka

Differential Revision: D48399818