rename to backend/interface.h (#285)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/285

this file is more about the interface between the backend and the executor runtime. The original name is `backend.h` and was renamed to `backend_registry.h` during refactor. But the backend registery is a really small part.

Techincally we can seperate out backend_registery, but it's just a few lines and feel like it might be overkill. Users also need to include two header files too and doesn't seem worth the effort

Reviewed By: JacobSzwejbka

Differential Revision: D48399818

fbshipit-source-id: f873e25d1cea12c457087b9011dfd5cbd3b2a47e