Adding support for delegate map in emitter (#334)
Summary:

LoweredBackendModules might optionally have a delegate debug handle map embedded in them if `preprocess` returned a result with a map inside it.

The emitter internally generates a mapping which is:
```
{'instruction_id' : {'name':delegate_name, 'debug_handle_map': delegate_identifier_map}
```
and returns this in the `EmitterOutput` object which is then subsequently copied over into the `ExecutorchProgram` object returned by `to_executorch()`. This enables us to serialize this map into etrecord which we can then use later for mapping back delegate internal operations to the original node in the graph.

Reviewed By: Jack-Khuu

Differential Revision: D49173918